Better (refactoring)
