+++ modules/constants.js
@@ -1,0 +1,20 @@
+export const InfuraEndpoints = {
+    ethereum: {
+        http: {
+            Main: 'https://mainnet.infura.io/mist',
+            Ropsten: 'https://ropsten.infura.io/mist',
+            Rinkeby: 'https://rinkeby.infura.io/mist',
+            Kovan: 'https://kovan.infura.io/mist'
+        },
+        websockets: {
+            Main: 'wss://mainnet.infura.io/ws',
+            Ropsten: 'wss://ropsten.infura.io/ws',
+            Rinkeby: 'wss://rinkeby.infura.io/ws',
+            Kovan: 'wss://kovan.infura.io/ws'
+        }
+    },
+    ipfs: {
+        gateway: 'https://ipfs.infura.io',
+        rpc: 'https://ipfs.infura.io:5001'
+    }
+}
